Fix: 로그인 여부에 따라 SSE 호출 여부 설정
