+++ src/hooks/useSSEMessage.ts
@@ -5,11 +5,14 @@ import { EventSourcePolyfill } from 'event-source-polyfill'
 import { authorizationInstance } from '@/api/instance'
 import { useAuthTokenStore } from '@/stores/auth-token'
 
-export function useSEEMessage() {
+export function useSSEMessage() {
   const [message, setMessage] = useState('')
   const { authToken } = useAuthTokenStore.getState()
+  const isLoggedIn = useAuthTokenStore.getState().isLoggedIn()
 
   useEffect(() => {
+    if (!isLoggedIn) return () => {}
+
     const eventSource = new EventSourcePolyfill(
       `${import.meta.env.VITE_BASE_URL}/api/alarm`,
       {
@@ -27,7 +30,7 @@ export function useSEEMessage() {
       authorizationInstance.post('/api/alarm/disconnect', {})
       eventSource.close()
     }
-  }, [authToken])
+  }, [authToken, isLoggedIn])
 
   return { message }
 }
